refactor: move ImportPositionEntity to webapps schema
