+++ operate/importer/src/main/java/io/camunda/operate/zeebeimport/ImportPositionHolder.java
@@ -7,9 +7,9 @@
  */
 package io.camunda.operate.zeebeimport;
 
-import io.camunda.operate.entities.meta.ImportPositionEntity;
 import io.camunda.operate.property.OperateProperties;
 import io.camunda.operate.store.ImportStore;
+import io.camunda.webapps.schema.entities.operate.ImportPositionEntity;
 import jakarta.annotation.PostConstruct;
 import java.io.IOException;
 import java.time.OffsetDateTime;
