+++ operate/importer/src/main/java/io/camunda/operate/zeebeimport/opensearch/OpensearchRecordsReader.java
@@ -24,7 +24,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import io.camunda.operate.Metrics;
 import io.camunda.operate.conditions.OpensearchCondition;
 import io.camunda.operate.entities.HitEntity;
-import io.camunda.operate.entities.meta.ImportPositionEntity;
 import io.camunda.operate.exceptions.NoSuchIndexException;
 import io.camunda.operate.exceptions.OperateRuntimeException;
 import io.camunda.operate.property.OperateProperties;
@@ -39,6 +38,7 @@ import io.camunda.operate.zeebeimport.ImportJob;
 import io.camunda.operate.zeebeimport.ImportListener;
 import io.camunda.operate.zeebeimport.ImportPositionHolder;
 import io.camunda.operate.zeebeimport.RecordsReader;
+import io.camunda.webapps.schema.entities.operate.ImportPositionEntity;
 import jakarta.annotation.PostConstruct;
 import java.io.IOException;
 import java.io.StringWriter;
