+++ operate/schema/src/main/java/io/camunda/operate/store/opensearch/OpensearchImportStore.java
@@ -15,13 +15,13 @@ import static io.camunda.operate.store.opensearch.dsl.RequestDSL.searchRequestBu
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.camunda.operate.Metrics;
 import io.camunda.operate.conditions.OpensearchCondition;
-import io.camunda.operate.entities.meta.ImportPositionEntity;
 import io.camunda.operate.property.OperateProperties;
 import io.camunda.operate.schema.IndexMapping;
 import io.camunda.operate.schema.indices.ImportPositionIndex;
 import io.camunda.operate.store.ImportStore;
 import io.camunda.operate.store.opensearch.client.sync.RichOpenSearchClient;
 import io.camunda.operate.util.Either;
+import io.camunda.webapps.schema.entities.operate.ImportPositionEntity;
 import java.io.IOException;
 import java.util.List;
 import java.util.Map;
@@ -110,6 +110,32 @@ public class OpensearchImportStore implements ImportStore {
     }
   }
 
+  @Override
+  public boolean getConcurrencyMode() {
+    final String indexName = importPositionType.getFullQualifiedName();
+    final Map<String, IndexMapping> indexMappings =
+        richOpenSearchClient.index().getIndexMappings(indexName);
+    if (indexMappings.get(indexName).getMetaProperties() == null) {
+      return false;
+    } else {
+      final Object concurrencyMode =
+          indexMappings.get(indexName).getMetaProperties().get(META_CONCURRENCY_MODE);
+      return concurrencyMode != null && (boolean) concurrencyMode;
+    }
+  }
+
+  @Override
+  public void setConcurrencyMode(final boolean concurrencyMode) {
+    final String indexName = importPositionType.getFullQualifiedName();
+    LOGGER.debug("Meta field will be updated. Index name: {}. ");
+    final PutMappingRequest request =
+        new PutMappingRequest.Builder()
+            .index(indexName)
+            .meta(META_CONCURRENCY_MODE, JsonData.of(concurrencyMode))
+            .build();
+    richOpenSearchClient.index().putMapping(request);
+  }
+
   private void withImportPositionTimer(final Callable<Void> action) throws Exception {
     metrics.getTimer(Metrics.TIMER_NAME_IMPORT_POSITION_UPDATE).recordCallable(action);
   }
@@ -127,32 +153,6 @@ public class OpensearchImportStore implements ImportStore {
                           .id(position.getId())
                           .upsert(entityProducer.apply(position))
                           .document(entityProducer.apply(position))));
-    }
-  }
-
-  @Override
-  public void setConcurrencyMode(final boolean concurrencyMode) {
-    final String indexName = importPositionType.getFullQualifiedName();
-    LOGGER.debug("Meta field will be updated. Index name: {}. ");
-    final PutMappingRequest request =
-        new PutMappingRequest.Builder()
-            .index(indexName)
-            .meta(META_CONCURRENCY_MODE, JsonData.of(concurrencyMode))
-            .build();
-    richOpenSearchClient.index().putMapping(request);
-  }
-
-  @Override
-  public boolean getConcurrencyMode() {
-    final String indexName = importPositionType.getFullQualifiedName();
-    final Map<String, IndexMapping> indexMappings =
-        richOpenSearchClient.index().getIndexMappings(indexName);
-    if (indexMappings.get(indexName).getMetaProperties() == null) {
-      return false;
-    } else {
-      final Object concurrencyMode =
-          indexMappings.get(indexName).getMetaProperties().get(META_CONCURRENCY_MODE);
-      return concurrencyMode != null && (boolean) concurrencyMode;
     }
   }
